requirements.txt
PyQt6
PyOpenGL